Move none page components into the components folder
