+++ packages/core/admin/admin/src/components/LanguageProvider/reducer.js
