+++ packages/core/admin/admin/src/components/LocaleToggle/reducer.js
