+++ packages/core/admin/admin/src/components/LocaleToggle/tests/reducer.test.js
