+++ packages/core/admin/admin/src/components/Onboarding/Li.js
