+++ packages/core/admin/admin/src/components/Onboarding/StaticLinks/StyledLink.js
