+++ packages/core/admin/admin/src/components/Onboarding/StaticLinks/index.js
