+++ packages/core/admin/admin/src/components/Onboarding/Video.js
