+++ packages/core/admin/admin/src/components/Onboarding/index.js
