+++ packages/core/admin/admin/src/components/Onboarding/reducer.js
