+++ packages/core/admin/admin/src/components/Onboarding/utils/formatAndStoreVideoArray.js
