+++ packages/core/admin/admin/src/components/PermissionsManager/actions.js
