+++ packages/core/admin/admin/src/components/PermissionsManager/constants.js
