+++ packages/core/admin/admin/src/components/PermissionsManager/index.js
