+++ packages/core/admin/admin/src/components/PermissionsManager/reducer.js
