+++ packages/core/admin/admin/src/containers/Admin/index.js
@@ -31,8 +31,8 @@ import InstalledPluginsPage from '../InstalledPluginsPage';
 import HomePage from '../HomePage';
 import MarketplacePage from '../MarketplacePage';
 import NotFoundPage from '../NotFoundPage';
-import OnboardingVideos from '../Onboarding';
-import PermissionsManager from '../PermissionsManager';
+import OnboardingVideos from '../../components/Onboarding';
+import PermissionsManager from '../../components/PermissionsManager';
 import PluginDispatcher from '../PluginDispatcher';
 import ProfilePage from '../ProfilePage';
 import SettingsPage from '../SettingsPage';
