+++ packages/core/admin/admin/src/reducers/index.js
@@ -1,7 +1,7 @@
 import globalReducer from '../containers/App/reducer';
 import adminReducer from '../containers/Admin/reducer';
-import languageProviderReducer from '../containers/LanguageProvider/reducer';
-import permissionsManagerReducer from '../containers/PermissionsManager/reducer';
+import languageProviderReducer from '../components/LanguageProvider/reducer';
+import permissionsManagerReducer from '../components/PermissionsManager/reducer';
 import menuReducer from '../containers/LeftMenu/reducer';
 
 // TODO move containers reducers into this folder
